move RepositoryForDoctrineProvider logic to abstract class
