+++ src/Contract/Bridge/RepositoryForDoctrineEntityProviderInterface.php
@@ -5,4 +5,9 @@ namespace Rector\Contract\Bridge;
 interface RepositoryForDoctrineEntityProviderInterface
 {
     public function provideRepositoryForEntity(string $name): ?string;
+
+    /**
+     * @return string[]
+     */
+    public function provideRepositoriesForEntities(): array;
 }
